Move substitution and regenerate docs
