+++ effekt/shared/src/main/scala/effekt/machine/Tree.scala
@@ -95,6 +95,7 @@ case class Clause(parameters: Environment, body: Statement)
  *     │─ [[ LiteralDouble ]]
  *     │─ [[ LiteralUTF8String ]]
  *     │─ [[ LiteralEvidence ]]
+ *     │─ [[ Hole ]]
  *
  * --------------------------------------------------
  */
